feat(.github/workflows/build-push-lodestar.yml): capture commit hash before deploy to avoid race conditions
This commit introduces a new job, `get-commit-hash`, that runs before the `deploy` job.
The purpose of this job is to capture the exact commit hash that will be deployed.
This is important to handle scenarios where a new commit is pushed to the remote repository between the `deploy` and `manifest` actions.
By capturing the commit hash before the `deploy` job, we ensure that the `manifest` job uses the correct commit hash, even if a new commit has been pushed in the meantime.